Cleanup, breaks item operations view
'Use' selection is not drawn and the row contains garbage chars at end.
Additionally warnings are generated from struct actor parameter.
